+++ checkout/src/main/java/com/mercadopago/sdk/android/checkout/presentation/extensions/NumberExtensions.kt
@@ -18,3 +18,17 @@ internal fun Float.toDecimalPartString(): String {
         .toInt()
     return cents.toString().padStart(2, '0')
 }
+
+internal fun BigDecimal.getTotal(): String {
+    val scaled = setScale(2, RoundingMode.HALF_UP)
+    return scaled.toBigInteger().toString()
+}
+
+internal fun BigDecimal.getTotalDecimalPart(): String {
+    val scaled = setScale(2, RoundingMode.HALF_UP)
+    val cents = scaled.remainder(BigDecimal.ONE)
+        .movePointRight(2)
+        .abs()
+        .toInt()
+    return cents.toString().padStart(2, '0')
+}
